requirements.txt
google-generativeai>=0.3.0
langchain>=0.1.0  # Only used for text splitting
